feat: SaveManager

+++ core/include/core/component/aura_damage.hpp
@@ -4,7 +4,7 @@ namespace core {
 
 struct AuraDamageComponent {
     int damage = 0;
-    float interval = 1000000.f;
+    int interval = 1000000;
 };
 
 }  // namespace core
